Update COT004_Suppress_quotation_page_if_equal_to_first_page_of_article.cs
Parentheses around the citation pages can now be defined as literary elements within the component or as part of the prefix/suffix of the citation pages. In both cases they should be suppressed correctly if necessary, i.e. if the article pages are omitted due to the condition "first citation page corresponds to first page of article".